UT-206: Improves footer and moves the new button to under settings.
feature: adds a button to the footer that exits whichever mode you are in back to OperationMode.Control always.
refactor: moves the new hiding button to be under settings, insted of the topbar.
chore: updated color of footer text to better represent a warning, rather than an error. (orangeish color vs redish color)
